refactor: Move data structure functions into its own module
