+++ src/utils/data.js
@@ -1,0 +1,11 @@
+export const arrayToMap = (csets) => {
+  const newCsets = {};
+  csets.forEach((cset) => {
+    newCsets[cset.node] = cset;
+  });
+  return newCsets;
+};
+
+export const mapToArray = csets => (
+  Object.keys(csets).map(node => csets[node])
+);
